+++ src/main/java/net/byteseek/incubator/matcher/automata/NfaMatcher.java
@@ -61,7 +61,7 @@ public final class NfaMatcher implements AutomataMatcher {
 	}
 
 	@Override
-	public long matches(WindowReader reader, long matchPosition, List<MatchResult> results) throws IOException {
+	public int matches(WindowReader reader, long matchPosition, List<MatchResult> results) throws IOException {
 		return 0; //TODO: implement.
 	}
 
@@ -118,7 +118,7 @@ public final class NfaMatcher implements AutomataMatcher {
 	}
 
 	@Override
-	public long matches(byte[] bytes, int matchPosition, List<MatchResult> results) {
+	public int matches(byte[] bytes, int matchPosition, List<MatchResult> results) {
 		return 0; //TODO: implement.
 	}
