+++ src/main/java/net/byteseek/incubator/matcher/multisequence/ListMultiSequenceMatcher.java
@@ -336,7 +336,7 @@ public final class ListMultiSequenceMatcher implements MultiSequenceMatcher {
 
 
     @Override
-    public long matches(WindowReader reader, long matchPosition, List<MatchResult> results) throws IOException {
+    public int matches(WindowReader reader, long matchPosition, List<MatchResult> results) throws IOException {
         return 0; //TODO: implement.
     }
 
@@ -356,7 +356,7 @@ public final class ListMultiSequenceMatcher implements MultiSequenceMatcher {
     }
 
     @Override
-    public long matches(byte[] bytes, int matchPosition, List<MatchResult> results) {
+    public int matches(byte[] bytes, int matchPosition, List<MatchResult> results) {
         return 0; //TODO: implement.
     }
